requirements.txt
Django>=5.2.7,<5.3
djangorestframework>=3.16.1,<3.17
psycopg2==2.9.9
python-dotenv>=1.0.0,<2.0.0
gunicorn>=21.2.0,<22.0.0
drf-spectacular>=0.28.0,<0.29.0
dj-database-url>=3.0.1,<3.1.0
#celery>=5.4.7,<5.5.0